Move errors.go from `runtime` to `node` package
